chapter 4

+++ src/error.rs
@@ -10,11 +10,11 @@ pub fn had_error() -> bool {
     HAD_ERROR.load(Ordering::Relaxed)
 }
 
-pub fn error(line: i32, message: String) {
+pub fn error(line: usize, message: String) {
     report(line, "".to_string(), message);
 }
 
-fn report(line: i32, error: String, message: String) {
+fn report(line: usize, error: String, message: String) {
     eprintln!("[line {line}] Error {error}: {message}");
     HAD_ERROR.store(true, Ordering::Relaxed);
 }
